@@ -1,4 +1,12 @@
+import 'package:edusign_v3/models/user_model.dart';
+import 'package:edusign_v3/services/storage_service.dart';
+
 class SignatureService {
-  static String defaultSignture =
+  static const String _defaultSignture =
       'iVBORw0KGgoAAAANSUhEUgAAAZAAAADICAYAAADGFbfiAAAGm0lEQVR4nO3dv26jSxkH4N8l+A7wJeQCOJIvISUdU9ATbgDcUVCw4gbOVFCyUG3FcXH6E4FEhUSQjmg3QggOUITCG9lxvs+biR2P/zyPZO1qs3HeNPllZt55vwQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAICzMv30AoBR0yQ/S/LXJA9J/vbpz4ckpVtVABytSZJfZhUWQ69Fr+IAOE7zrFYc217zPuUBcEyukvwuyf/yPCjuk/x+49++zXKVAsAFmiT5cZJvMrzCuE/yLstzkMXGx8qhiwWgv1mSL5N8zPj21GNwPP5/Zx8AF+pxtbHtbON9kl8k+WLjcxcb/292iIIB6OuLJH/M+BZVTXK95fNnsfoAuCizJL/N+EqjvOA9Jnl+PvKSzwPgBJUkX2V4tTFP2w3yuvEet3urEoCjsO18Y5HtW1Rjbgbea7Z7qQAcg6ssu6mGtqnqp4+/Rhl4v/lOlQJwFEqGt6lus1w57HLB72rkfQE4Udu2qWr2s710led3Q+5j8i7ASZom+U2e/2C/y+6rjXVXGQ6n126DAdDBNE9HqO/jUHyboXZdLbsAJ2KS5IcZn0v1dd6mC2qS5WDEoVEmAByxx06qobOHmuUq4C2n3v4lz8OjvuHXA2BHJcOdVO+z3KI6xKj0OvD1bw/0tQFoMM3w2cZd3n6lsankeXj868A1APAZswxf+Kvpc7t7NlDLfXRcARyFacZXG/P0u1sxzfBzQEqnegDIcvvnJ9l+ttHTWLvuvGNNABdtrJPqLn1XG5tqhoMNgAMaGy3ynyQfcnyTa+fRcQXQ1bbVxj5Hi+zTdRyaA3Sx7Zb4XQ7fgttiaEDiQ45vhQRwVqZZdlIN/QC+zfF3Lk0yPE+rdKwJ4KxdZfx54ouczm/vixhTAnAQswy34D52K816FfYKNcPhB8AezTIeHDXH04b7UiU6rgDe1Czj49NrTi84kvExJcIDYA9m2b5VNe1V2I7GHkmrXRdgR9OMB8cip3XGsWloTInwANjRNMPTcM8hOB4tIjwA9maa8eC4S//hhvtSIzwA9mKS5QXAseAovQp7AzcRHgA7ewyOoZvj9zm/seUlwgNgZz9N8t+MB8e5tbFudlwJD4BGJcPzns41OJLnTxUUHgANSoaD4/ES4DkGR/K8Xfc2p3tvBeCgSsaD4x9JftCtssN4H+NJAJqUjAfHpfwmLjwAGkwzfnv8ErasHtWsvt+vc/7fL8BObjLckrt+eHzTrbrDqVl9zx/6lgJw3CYZf6DT+hbOJXQe1TxdaQEw4irbzzoekrzrVt1hLSI8AF6k5PNbVrNOtR3SZqvuJWzTAbxazfZVx/tcxsHxZniUrtUAHLk/ZHt4XMpv4MIDoEGNLatEeAC82CTb73dcwsXAR8ID4IWu8vzxq+uvP+UyzjsS4QHwYtcZ77T6Nuc/y2qd8AB4oZrLvlG+TngAvNDYltWltOeuEx4ADYZWHaVnQZ0ID4BG91n90PwulzHHatPmmJbStRqAE/G9LGc7LT79/dKsP8P8UldfADQqeRoel7j6AqDRr7LasrrL5VyOBGAH86zC48+5vG4zAF5hHk8RBKDRu3gQFACNaoQHAI1qhAcAjWqEBwCNaoQHAI1qhAcAjWqEBwCNaoQHAI1qhAcAjWqEBwCNaoQHAI1qhAcAjWqEBwCNaoQHAI1qhAcAjWqEBwCNaoQHAI1qVuHxrm8pAJyK9ScJzvuWAsCpEB4ANLvJKjxK31IAOBUlwgOARiWr8LjpWwoAp6LEygOARiXCA4BGP4rwAKDRdZJ/RngA0OA6yccIDwAalAgPABqVOPMAoFGJex4ANCqx8gCgUYmVBwCNSkzVBaBRiW0rABqV2LYCoFGJZ5gD0KhEeADQqER4ANCoRHgA0KhEeADQqER4ANDo5xEeADS6ySo8ft25FgBOxHp4fOhcCwAnYh7bVgA0qhEeADSqER4ANKoRHgA0qlmFx23fUgA4BZMkX+ZpeEy6VgTA0Zsk+SbCA4AGkyR/j/AAoNEiq/D4d5Jpz2IAOA01q/D4Lsn3u1YDwEmoWYXHQ5JZz2IAOH6TJF/laXiUngUBcPyu8rTbSngA8FmzJB8jPABodB/hAcArCA8AXuU6yzsf153rAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA4b/8H321kFQhvp4IAAAAASUVORK5CYII=';
+
+  static Future<String> getSignature(User user) async {
+    String? signature = await StorageService.read(key: 'signature_${user.username}');
+    return signature ?? _defaultSignture;
+  }
 }
